requirements.txt
flask
flask-restful
flask-sqlalchemy
redis
python-dotenv
pytest